Moved integration with thunder-registration into a package extension
