+++ coregistration/eolearn/coregistration/__init__.py
@@ -2,12 +2,6 @@
 A collection of tools and EOTasks for image co-registration
 """
 
-from .coregistration import (
-    ECCRegistrationTask,
-    InterpolationType,
-    PointBasedRegistrationTask,
-    RegistrationTask,
-    ThunderRegistrationTask,
-)
+from .coregistration import ECCRegistrationTask, InterpolationType, PointBasedRegistrationTask, RegistrationTask
 
 __version__ = "1.1.0"
