+++ coregistration/eolearn/coregistration/extra/__init__.py
@@ -1,0 +1,3 @@
+"""
+A subfolder containing integrations with other packages
+"""
