+++ coregistration/requirements-thunder.txt
@@ -1,0 +1,1 @@
+thunder-registration
