+++ coregistration/requirements.txt
@@ -1,3 +1,2 @@
 eo-learn-core
 opencv-contrib-python-headless>=4.1
-thunder-registration
