+++ coregistration/setup.py
@@ -40,6 +40,7 @@ setup(
     packages=find_packages(),
     include_package_data=True,
     install_requires=parse_requirements("requirements.txt"),
+    extras_require={"THUNDER": parse_requirements("requirements-thunder.txt")},
     zip_safe=False,
     classifiers=[
         "Development Status :: 5 - Production/Stable",
